pyproject.toml
[project]
name = "knowledge-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.5.1",
    "inquirer>=3.4.0",
    "lancedb>=0.22.0",
    "openai>=1.79.0",
    "pandas>=2.2.3",
    "pylance>=0.27.2",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "tantivy>=0.24.0",
]
